feat: Update Geruest server configuration and enhance language support
- Changed default server port from 80 to 8080 for easier access on Linux.
- Added thread pool configuration options to optimize server performance.
- Implemented language configuration with support for multiple languages.
- Enhanced server shutdown process to ensure graceful termination of worker threads.
- Updated ContentBuilder and HTMLBuilder to support language-specific file handling.
- Improved error handling in Handler for socket communication.
- Added methods in ServerData to manage available languages and default language.

+++ src/handler/Handler.hpp
@@ -61,7 +61,7 @@ private:
 	const std::string IP;
 
 	char *buffer;
-	int bufferLength = 0;
+	ssize_t bufferLength = 0;
 
 	bool readSocket() { return readSocket(buffer, BUFFER_SIZE); }
